Improve Flyout arrow frame strata behavior
Originally the Flyout arrow was always on FULLSCREEN, but when one action bar was above another and both had Flyouts on neighboring buttons, the arrow of the top button would show above the flyout of the bottom one. Arrows are now shifted up to FULLSCREEN when the flyout menu shows and dropped back to the original strata when hidden.